Refactor responsabilidades HomeActivity
