+++ app/src/main/java/com/victor/dogbreeds/ui/home/HomeContract.kt
@@ -1,11 +1,14 @@
 package com.victor.dogbreeds.ui.home
 
 import com.victor.dogbreeds.business.models.BreedsModel
+import com.victor.dogbreeds.business.models.UserModel
 import com.victor.dogbreeds.util.ConnectionType
 
 interface HomeContract {
     interface View {
         fun setBreeds(breeds: List<BreedsModel>)
+        fun showErrorToastMessage()
+        fun setUserModel(userModel: UserModel)
     }
 
     interface Presenter {
